feat: modal component with uplifting state to show/hide
